refactor: post widget 분리
