+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { BrowserRouter as Router } from "react-router-dom"
-import Header from "./widgets/ui/Header.tsx"
-import Footer from "./widgets/ui/Footer.tsx"
+import Header from "./shared/ui/Header.tsx"
+import Footer from "./shared/ui/Footer.tsx"
 import PostsManagerPage from "./pages/postManager/PostManagerPage.tsx"
 
 const App = () => {
