+++ src/entity/post/ui/index.ts
@@ -1,0 +1,4 @@
+export * from "./PostTable"
+export * from "./PostHeader"
+export * from "./PostEditForm"
+export * from "./PostAddForm"
