+++ src/pages/postManager/PostManagerPage.tsx
@@ -1,8 +1,12 @@
-import { PostHeader, PostAddDialog, PostEditDialog, PostDetailDialog } from "@/widgets/post/ui"
-import { UserDetailDialog } from "@/widgets/user/ui/UserDetailDialog"
 import { CardLayout } from "@/app/layouts/CardLayout"
-import { PostList } from "@/widgets/post/ui/PostList"
 
+import { PostList } from "@/widgets/post/ui/PostList"
+import { PostAddDialog, PostEditDialog, PostDetailDialog } from "@/widgets/post/ui"
+import { UserDetailDialog } from "@/widgets/user/ui/UserDetailDialog"
+
+import { PostHeader } from "@/entity/post/ui"
+
+// dialog가 많아지면 페이지에서 빼버리는것도 고려해볼것
 const PostManagerPage = () => {
   return (
     <CardLayout header={<PostHeader />}>
