+++ src/shared/ui/index.ts
@@ -5,3 +5,8 @@ export * from "./Select"
 export * from "./Table"
 export * from "./Textarea"
 export * from "./Card"
+export * from "./Loading"
+export * from "./Pagination"
+export * from "./SearchInput"
+export * from "./SortControls"
+export * from "./TagFilter"
