+++ src/widgets/post/ui/index.ts
@@ -1,5 +1,4 @@
 export * from "./PostAddDialog"
+export * from "./PostAddDialog"
 export * from "./PostEditDialog"
 export * from "./PostDetailDialog"
-export * from "./PostHeader"
-export * from "./PostList"
